TEST: Split tests between modules and add doc
